Update Formalism.ipynb
In this notebook we present the functions used for finding and analayzing impefect GKP states in "Towards practical qubit computation using approximate error-correcting grid states" (I. Tzitrin, J. E. Bourassa, N. C. Menicucci, K. K Sabapathy; arXiv:1910.03673). For more details on the implementations here, please see the paper, particularly App. C.

In Section 2 of this notebook we set up useful functions; in Section 3 we generate two classes of normalizable GKP states; in Section 4 we use the modular subsystem decomposition to plot the location of the qubit associated with imperfect GKP states; in Section 5 we find approximations to GKP states, and in Section 6 we calculate several properties of these approximate states.

See the StatePrepCircuits notebook for optimization of GBS devices for GKP state preparation.